Refactor JWT and JWT to derive from a common parent class
